update: ファイル名変更
認証のauth_indexをindexに、メインメニューのindexをdashboardに変更
